Store our frame time as int64_t, the same type lv2 functions take

+++ src/effects/lv2/LV2Effect.h
@@ -243,7 +243,7 @@ private:
 
    // Position info
    float mPositionSpeed{ 1.0f };
-   float mPositionFrame{ 0.0f };
+   int64_t mPositionFrame{ 0 };
 
    double mLength{};
